Pig Latin
A fun exercise I did during my early years of programming, specifically in APCS. It translate any 5 letter sentence to its Pig Latin variant. Although there are some flaws, it was an important milestone in my CS career as it was the first bit of code that I truly felt proud of.